feat: OutputView 작성
